% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{panel_composition}
\alias{panel_composition}
\title{Select and renormalise a measured glycoform panel}
\usage{
panel_composition(composition, panel)
}
\arguments{
\item{composition}{Named full composition.}

\item{panel}{Character or integer panel specification.}
}
\value{
Named within-panel-normalised composition.
}
\description{
Select and renormalise a measured glycoform panel
}
