% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.glyco_panel_audit}
\alias{plot.glyco_panel_audit}
\alias{plot.glyco_robustness_audit}
\alias{plot.glyco_panel_set}
\alias{plot.glyco_pathway_fit}
\title{Plot a panel-design audit}
\usage{
\method{plot}{glyco_panel_audit}(x, ...)

\method{plot}{glyco_robustness_audit}(x, ...)

\method{plot}{glyco_panel_set}(x, ...)

\method{plot}{glyco_pathway_fit}(x, ...)
}
\arguments{
\item{x}{Object to plot.}

\item{...}{Additional graphical parameters.}
}
\description{
Plot a panel-design audit
}
