% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.glyco_pathway}
\alias{plot.glyco_pathway}
\title{Plot a biosynthetic pathway}
\usage{
\method{plot}{glyco_pathway}(x, show_rate_labels = TRUE, ...)
}
\arguments{
\item{x}{A \code{glyco_pathway} object.}

\item{show_rate_labels}{Show effective rate-class labels on reaction edges.}

\item{...}{Additional graphical parameters passed to the graph plot.}
}
\description{
Plot a biosynthetic pathway
}
