% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot the local sensitivity matrix}
\usage{
plot_sensitivity(audit, ...)
}
\arguments{
\item{audit}{A \code{glyco_panel_audit} object.}

\item{...}{Additional graphical parameters passed to \code{graphics::image}.}
}
\value{
The input invisibly.
}
\description{
Plot the local sensitivity matrix
}
