% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Launch the glycoPathDesign Shiny application}
\usage{
run_app(launch.browser = interactive(), ...)
}
\arguments{
\item{launch.browser}{Passed to \code{shiny::runApp}.}

\item{...}{Additional arguments passed to \code{shiny::runApp}.}
}
\description{
Launch the glycoPathDesign Shiny application
}
