% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{validate_pathway}
\alias{validate_pathway}
\title{Validate a glycoform biosynthetic pathway model}
\usage{
validate_pathway(pathway)
}
\arguments{
\item{pathway}{A \code{glyco_pathway} object.}
}
\value{
The input invisibly, or an error when invalid.
}
\description{
Validate a glycoform biosynthetic pathway model
}
