% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{write_pathway}
\alias{write_pathway}
\title{Write a pathway definition to a directory}
\usage{
write_pathway(pathway, directory)
}
\arguments{
\item{pathway}{A \code{glyco_pathway} object.}

\item{directory}{Output directory.}
}
\value{
The normalised directory path invisibly.
}
\description{
Write a pathway definition to a directory
}
